✅ Rewrite tests dealing with tuple
